move: airport, airline 패키지 분리
