--- flight-service/src/main/java/com/oneul_tanda/flight_service/application/dtos/AirportCommand.java
@@ -1,16 +1,0 @@
-package com.oneul_tanda.flight_service.application.dtos;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class AirportCommand {
-
-    private String code;
-    private String name;
-    private String city;
-    private String country;
-}
