+++ flight-service/src/main/java/com/oneul_tanda/flight_service/application/service/AirportExternalService.java
@@ -3,9 +3,9 @@ package com.oneul_tanda.flight_service.application.service;
 import com.amadeus.Amadeus;
 import com.amadeus.Params;
 import com.amadeus.resources.Location;
-import com.oneul_tanda.flight_service.presentation.dtos.AirportResponse;
+import com.oneul_tanda.flight_service.presentation.dtos.airport.AirportResponse;
 import com.oneul_tanda.flight_service.domain.entity.Airport;
-import com.oneul_tanda.flight_service.domain.repository.AirportRepository;
+import com.oneul_tanda.flight_service.domain.repository.airport.AirportRepository;
 import java.util.Arrays;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
