--- flight-service/src/main/java/com/oneul_tanda/flight_service/domain/repository/AirlineRepository.java
@@ -1,4 +1,0 @@
-package com.oneul_tanda.flight_service.domain.repository;
-
-public interface AirlineRepository{
-}
