+++ flight-service/src/main/java/com/oneul_tanda/flight_service/domain/repository/airline/AirlineRepository.java
@@ -1,0 +1,4 @@
+package com.oneul_tanda.flight_service.domain.repository.airline;
+
+public interface AirlineRepository{
+}
