+++ flight-service/src/main/java/com/oneul_tanda/flight_service/infrastructure/repository/airline/AirlineRepositoryImpl.java
@@ -1,7 +1,7 @@
-package com.oneul_tanda.flight_service.infrastructure.repository;
+package com.oneul_tanda.flight_service.infrastructure.repository.airline;
 
 import com.oneul_tanda.flight_service.domain.entity.AirLine;
-import com.oneul_tanda.flight_service.domain.repository.AirlineRepository;
+import com.oneul_tanda.flight_service.domain.repository.airline.AirlineRepository;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
