+++ flight-service/src/main/java/com/oneul_tanda/flight_service/infrastructure/repository/airport/AirportJpaRepository.java
@@ -1,7 +1,7 @@
-package com.oneul_tanda.flight_service.infrastructure.repository;
+package com.oneul_tanda.flight_service.infrastructure.repository.airport;
 
 import com.oneul_tanda.flight_service.domain.entity.Airport;
-import com.oneul_tanda.flight_service.domain.repository.AirportRepository;
+import com.oneul_tanda.flight_service.domain.repository.airport.AirportRepository;
 import java.util.UUID;
 import org.springframework.data.jpa.repository.JpaRepository;
