+++ flight-service/src/main/java/com/oneul_tanda/flight_service/presentation/dtos/airport/AirportResponse.java
@@ -1,4 +1,4 @@
-package com.oneul_tanda.flight_service.presentation.dtos;
+package com.oneul_tanda.flight_service.presentation.dtos.airport;
 
 import com.oneul_tanda.flight_service.domain.entity.Airport;
 import java.util.UUID;
